requirements.txt
databricks-sdk>=0.38.0
openai>=1.50.0
pyyaml>=6.0
tabulate>=0.9.0
